+++ src/constants.ts
@@ -1,13 +1,17 @@
 export const CONSTANTS = {
 	integrationExtension: '.integration.json',
 	scenariosDirectoryName: 'scenarios',
-    scenarioInputFileName: 'input.txt',
+	importsDirectoryName: 'imports',
 
 	panelTitlePrefix: 'Preview: ',
 	statusbarTitlePrefix: 'Scenario: ',
 	
 	configKeyEndpointUrl: 'stitch.endpointUrl',
 	previewActiveContextKey: 'stitch:previewActive'
+};
+
+export const MESSAGES = {
+	endpointUrlNotConfigured: `The '${CONSTANTS.configKeyEndpointUrl}' is not configured, please set this up in File -> Preferences -> Settings`
 };
 
 export const COMMANDS = {